requirements.txt
numpy==2.4.2
Pillow==12.1.0
pyperclip==1.11.0
Requests==2.32.5
torch==2.9.1+cu130
